GUI: Major restructuring of the underlying interface page's XML.  XML structure for options now mirrors the option_doc path structure in the C++ files.  This removes the need for defining which files contain options for which classes (the optdoc text neatly contains that information already).
Restructuring of the interface page a bit, to reflect the multi-tiered menu of the option_docs.  Not quite perfect, but getting there.

This compartmenting of options should pave the way for replacing entire sections with more functional interfaces (ie, replacing the 'talents' optdoc section with a custom interface, instead of just a huge list of fields).

git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@2853 3b4652a1-8050-0410-ac47-3d40261b0f8b