One small step towards better item caching.
These changes make use of the current url/xml cache. The difference is that the caches for the item from each possible site is checked before anything is downloaded.

i.e. the old method looked like pretty much this:

Step 1) See if the wowarmory item url is in the cache. If not, go to Step 2
Step 2) Try to download from the wowarmory. If it fails, go to Step 3
Step 3) See if the www.wowhead.com item url is in the cache. If not, go to Step 4
Step 4) Try to download from www.wowhead.com. If it fails go to Step 5
Step 5) See if the ptr.wowhead.com item url is in the cache. If not, go to Step 6
Step 6) Try to download from ptr.wowhead.com. If it fails go to Step 7
Step 7) See if the db.mmo-champion.com item url is in the cache. If not, go to Step 8
Step 8) Try to download from db.mmo-champion.com. If this fails, we return a failure.

The problem of course was if the item was only on db.mmo-champion.com we still had try to download the item from wowarmory, www.wowhead.com and ptr.wowhead.com before we could try to get the item from the db.mmo-champion.com cache.

The new procedure is this:

Step 1) See if the item is in the cache for wowarmory, www.wowhead.com, ptr.wowhead.com or db.mmo-champion.com. If not, go to Step 2
Step 2) Try to download from wowarmory (actual order depends on the function). If it fails go to Step 3
Step 3) Try to download from www.wowhead.com. If it fails go to Step 4
Step 4) Try to download from ptr.wowhead.com. If it fails go to Step 5
Step 5) Try to download from db.mmo-champion.com. If that fails, return a failure.

This still isn't a proper item cache which just caches the item information. 
We're still basically caching entire webpages which is somewhat...redundant.

Still, it's an improvement over the current situation.

git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@2987 3b4652a1-8050-0410-ac47-3d40261b0f8b