First (major) part of adding racial support to SimCraft and paying more attention to the value of stats etc. Also adds VE tracking for Shadow Priests and little things like the hidden +1 spi on Lightweave Embroidery and Darkglow Embroidery.
Each class has been assigned a default race if it's not overridden either as an option or set by the armory/wowhead etc. The default class was chosen as one that didn't have racials that modified stats (i.e. no Gnomes, Humans or Taurens).

A lot of the values for stats at different levels and for different races etc. were moved into sc_rating.cpp with methods added to access them.

The goal for the stat stuff is that the stats seen in the report should match what the player would see in-game when they had no buffs on and are in caster form (i.e. not in cat form or whatever). I would say it should match what the Armory says (and in general it does) except that the Armory has bugs of it own. ;P

I've not finished checking all of the classes yet to ensure that all of their talents that effect stats etc. are correct. I've only done Priests and Druids so far.

It's highly likely that I'll need to make further changes to the other classes to bring them in line.

Also, I've not yet added any support for the other passive racials (other than the Tauren health one, the Human spirit one and the gnome intellect one) nor any of the active racials.

A number of the .simcraft profiles will probably need to be regenerated to include the: race=<whatever> line so that the race on wowhead matches that of the .simcraft profile. (BTW, wowhead's own "stats" summary is fairly bad and shouldn't be trusted.)

git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@3135 3b4652a1-8050-0410-ac47-3d40261b0f8b