More changes due to race support and stat verification.
Also added in the expertise/crit/beast slaying passive racials.
This includes Heroic Presence which will be turned on for anyone in a party with a Draenei (or if overrides.heroic_presence=1)

Added Arcane Fortitude for Mages and a Glyph.

Still a few more classes I need to check to confirm that their stats match with what the Armory/Game says.

Also still need to add the active racials, Blood Fury and Berserking.

As an aside: can whomever it is that looks after the Hunter module, can you let me know what line 1554 is from, i.e.
    base_multiplier *= 1.05;



git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@3138 3b4652a1-8050-0410-ac47-3d40261b0f8b