Generalise DoT haste scaling.
The sim checks "int scale_ticks_with_haste()" to determine it. If it's 0, then it won't modify tick time with haste, otherwise it will. Default is 0.
Overload this method with the spells that tick faster with haste.

Also renamed the "haste_boost" option with SW:P to "haste_gain_percentage>" and made it into a general option.

git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@3586 3b4652a1-8050-0410-ac47-3d40261b0f8b