Added support to determine if an item is Heroic or not. It will parse mmo-champion, wowhead and armory tooltips looking for the Heroic tag. Profiles now have the optional field in gear of heroic=<0|1> (default is 0). Internally can use the method item_t::heroic() to check if the given item is heroic or not.
As part of this change I've updated the current trinkets/unique items that had differences between the non-heroic and heroic versions making use of this flag rather than (where available...and it certainly wasn't always) using things like item_id or checking stats etc.
(Also added a missing proc weapon).

This is the coding side of the fix.
The profiles (or at least those that use any Heroic items, or in particular those with procs, uses, etc.) need to be regenerated.

Also (as per some research from a 3rd-party) made some changes to how Tiny Abomination In A Jar works. As part of this change, there's a change to the melee stuff of the Enhancement Shaman. It's a change I'm not 100% sure of myself but it seems to fix up some crazy stuff that was happening.

git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@4394 3b4652a1-8050-0410-ac47-3d40261b0f8b