Added generate_profiles.simc
This is a meta-profile that includes all the commands to generate our current BiS profiles (for those generated via wowhead). Handy when we want to regenerate all profiles in a hurry.

Also added Death_Knight_T10_14_00_57.simc, added a talent spec and glyphs to Death_Knight_T10_00_17_54 and moved all the Death Knight T10 profiles to a new simcraftdkt10 wowhead user.

git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@4473 3b4652a1-8050-0410-ac47-3d40261b0f8b