Modified the following Paladin attacks to generate motes of anger for TAJ:
Hammer of Wrath
Seal of Justice proc and judgement
Seal of Light proc and judgement
Seal of Righteousness proc and judgement
Seal of Vengeance DOT and judgement
Seal of Wisdom proc and judgement
Exorcism

The manifest anger damage percentage increased by anywhere from 0.1% to 0.5% as a result of these changes. We are getting closer.


git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@4523 3b4652a1-8050-0410-ac47-3d40261b0f8b